requirements.txt
mistralai>=1.12.0
streamlit>=1.54.0
python-dotenv>=1.0.0
